WIP - Enables user to dynamically set connect back ip - useful if debugger is  on a remote instance (e.g. a vscode ssh session)
By settings the X-Forwarded-For, this will tell XDebug to connect back to
that ip vs the one the request actually came from. This trick can be used
to tell XDebug that it should connect to your remote VS Code instasnce which
is on a different IP than the computer you are accsesing the PHP site from.